+++ bin/commands/dnd.rb
@@ -22,6 +22,7 @@ command 'dnd' do |g|
   g.desc "Retrieves a user's current Do Not Disturb status."
   g.long_desc %( Retrieves a user's current Do Not Disturb status. )
   g.command 'info' do |c|
+    c.flag 'team_id', desc: 'Encoded team id where passed in user param belongs, required if org token is used. If no user param is passed, then a team which has access to the app should be passed.'
     c.flag 'user', desc: 'User to fetch status for (defaults to current user).'
     c.action do |_global_options, options, _args|
       puts JSON.dump($client.dnd_info(options))
@@ -41,6 +42,7 @@ command 'dnd' do |g|
   g.long_desc %( Retrieves the Do Not Disturb status for up to 50 users on a team. )
   g.command 'teamInfo' do |c|
     c.flag 'users', desc: 'Comma-separated list of users to fetch Do Not Disturb status for.'
+    c.flag 'team_id', desc: 'Encoded team id where passed in users belong, required if org token is used.'
     c.action do |_global_options, options, _args|
       puts JSON.dump($client.dnd_teamInfo(options))
     end