+++ lib/slack/web/api/endpoints/chat.rb
@@ -145,18 +145,18 @@ module Slack
           #
           # @option options [channel] :channel
           #   Channel, private group, or IM channel to send message to. Can be an encoded ID, or a name. See below for more details.
-          # @option options [Object] :text
-          #   How this field works and whether it is required depends on other fields you use in your API call. See below for more detail.
           # @option options [Object] :as_user
-          #   Pass true to post the message as the authed user, instead of as a bot. Defaults to false. See authorship below. This argument may not be used with newer bot tokens.
+          #   Pass true to post the message as the authed user, instead of as a bot. Defaults to false. See authorship below.
           # @option options [Object] :attachments
           #   A JSON-based array of structured attachments, presented as a URL-encoded string.
           # @option options [Object] :blocks
           #   A JSON-based array of structured blocks, presented as a URL-encoded string.
+          # @option options [Object] :draft_id
+          #   The id of the draft associated with the message.
           # @option options [Object] :icon_emoji
-          #   Emoji to use as the icon for this message. Overrides icon_url. See authorship below. Use with bot tokens requires chat:write.customize.
+          #   Emoji to use as the icon for this message. Overrides icon_url. Must be used in conjunction with as_user set to false, otherwise ignored. See authorship below.
           # @option options [Object] :icon_url
-          #   URL to an image to use as the icon for this message. See authorship below. Use with bot tokens requires chat:write.customize.
+          #   URL to an image to use as the icon for this message. Must be used in conjunction with as_user set to false, otherwise ignored. See authorship below.
           # @option options [Object] :link_names
           #   Find and link channel names and usernames.
           # @option options [Object] :mrkdwn
@@ -165,6 +165,8 @@ module Slack
           #   Change how messages are treated. Defaults to none. See below.
           # @option options [Object] :reply_broadcast
           #   Used in conjunction with thread_ts and indicates whether reply should be made visible to everyone in the channel or conversation. Defaults to false.
+          # @option options [Object] :text
+          #   How this field works and whether it is required depends on other fields you use in your API call. See below for more detail.
           # @option options [Object] :thread_ts
           #   Provide another message's ts value to make this message a reply. Avoid using a reply's ts value; use its parent instead.
           # @option options [Object] :unfurl_links
@@ -172,7 +174,7 @@ module Slack
           # @option options [Object] :unfurl_media
           #   Pass false to disable unfurling of media content.
           # @option options [Object] :username
-          #   Set your bot's user name. See authorship below. Use with bot tokens requires chat:write.customize.
+          #   Set your bot's user name. Must be used in conjunction with as_user set to false, otherwise ignored. See authorship below.
           # @see https://api.slack.com/methods/chat.postMessage
           # @see https://github.com/slack-ruby/slack-api-ref/blob/master/methods/chat/chat.postMessage.json
           def chat_postMessage(options = {})
@@ -238,8 +240,10 @@ module Slack
           #   Timestamp of the message to add unfurl behavior to.
           # @option options [Object] :unfurls
           #   URL-encoded JSON map with keys set to URLs featured in the the message, pointing to their unfurl blocks or message attachments.
+          # @option options [Object] :user_auth_blocks
+          #   Provide a JSON based array of structured blocks presented as URL-encoded string to send as an ephemeral message to the user as invitation to authenticate further and enable full unfurling behavior.
           # @option options [Object] :user_auth_message
-          #   Provide a simply-formatted string to send as an ephemeral message to the user as invitation to authenticate further and enable full unfurling behavior.
+          #   Provide a simply-formatted string to send as an ephemeral message to the user as invitation to authenticate further and enable full unfurling behavior. Provides two buttons, Not now or Never ask me again.
           # @option options [Object] :user_auth_required
           #   Set to true or 1 to indicate the user must install your Slack app to trigger unfurls for this domain.
           # @option options [Object] :user_auth_url