+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.16.1 (Next)
 
 * Your contribution here.
+* [#359](https://github.com/slack-ruby/slack-ruby-client/pull/359): Handle non-JSON 500 errors - [@agrobbin](https://github.com/agrobbin).
 
 ### 0.16.0 (2021/01/24)